requirements.txt
#pip install -r requirements.txt


django
pillow
django-admin-interface
django-crispy-forms
mysqlclient
pymysql
djangorestframework